+++ app/views/handlers/calculated_summary.py
@@ -1,17 +1,17 @@
-from app.views.contexts.calculated_summary import (
-    build_view_context_for_calculated_summary,
-)
+from app.views.contexts.calculated_summary_context import CalculatedSummaryContext
 from app.views.handlers.content import Content
 
 
 class CalculatedSummary(Content):
     def get_context(self):
-        return build_view_context_for_calculated_summary(
+        calculated_summary_context = CalculatedSummaryContext(
             self._language,
-            self._current_location,
             self._schema,
             self._questionnaire_store.answer_store,
-            self._questionnaire_store.progress_store,
             self._questionnaire_store.list_store,
+            self._questionnaire_store.progress_store,
             self._questionnaire_store.metadata,
+        )
+        return calculated_summary_context.build_view_context_for_calculated_summary(
+            self._current_location
         )
