+++ app/views/handlers/question.py
@@ -70,7 +70,7 @@ class Question(BlockHandler):
                 answer_store=self._questionnaire_store.answer_store,
                 list_store=self._questionnaire_store.list_store,
                 progress_store=self._questionnaire_store.progress_store,
-                metadata=self._questionnaire_store.metadata
+                metadata=self._questionnaire_store.metadata,
             )
 
             context["list"] = {
