+++ data-source/json/test_list_collector.json
@@ -316,7 +316,8 @@
                         },
                         {
                             "id": "group-summary",
-                            "type": "SectionSummary"
+                            "title": "List Collector Summary",
+                            "type": "ListCollectorSummary"
                         }
                     ]
                 }
