+++ data-source/jsonnet/england-wales/household/blocks/who-lives-here/who_lives_here_section_summary.jsonnet
@@ -1,4 +1,5 @@
 {
   id: 'who-lives-here-section-summary',
-  type: 'SectionSummary',
+  title: 'People who live here',
+  type: 'ListCollectorSummary',
 }
