+++ tests/app/views/contexts/test_hub_context.py
@@ -12,7 +12,7 @@ def router(schema, answer_store, list_store, progress_store):
 
 
 def test_get_not_started_row_for_section(
-    schema, progress_store, answer_store, list_store, router
+    schema, progress_store, answer_store, list_store
 ):
     expected = {
         "title": "Breakfast",
@@ -36,7 +36,7 @@ def test_get_not_started_row_for_section(
         list_store=list_store,
         schema=schema,
         answer_store=answer_store,
-        metadata={}
+        metadata={},
     )
 
     actual = hub.get_row_context_for_section(
@@ -49,7 +49,7 @@ def test_get_not_started_row_for_section(
 
 
 def test_get_completed_row_for_section(
-    schema, progress_store, answer_store, list_store, router
+    schema, progress_store, answer_store, list_store
 ):
     expected = {
         "title": "Breakfast",
@@ -74,7 +74,7 @@ def test_get_completed_row_for_section(
         list_store=list_store,
         schema=schema,
         answer_store=answer_store,
-        metadata={}
+        metadata={},
     )
 
     actual = hub.get_row_context_for_section(
@@ -103,5 +103,6 @@ def test_get_context(schema, progress_store, answer_store, list_store, router):
         "submit_button": "Continue",
     }
 
-    assert expected_context == hub.get_context(survey_complete=False,
-                                               enabled_section_ids=router.enabled_section_ids)
+    assert expected_context == hub.get_context(
+        survey_complete=False, enabled_section_ids=router.enabled_section_ids
+    )
