+++ tests/app/views/contexts/test_summary_context.py
@@ -8,10 +8,10 @@ from app.data_model.progress_store import ProgressStore
 from app.questionnaire.location import Location
 from app.questionnaire.questionnaire_schema import DEFAULT_LANGUAGE_CODE
 from app.utilities.schema import load_schema_from_name
-from app.views.contexts.calculated_summary import (
-    build_view_context_for_calculated_summary,
+from app.views.contexts.calculated_summary_context import CalculatedSummaryContext
+from app.views.contexts.list_collector_summary_context import (
+    ListCollectorSummaryContext,
 )
-from app.views.contexts.list_collector_summary_context import ListCollectorSummaryContext
 from app.views.contexts.summary_context import SummaryContext
 from tests.app.app_context_test_case import AppContextTestCase
 
@@ -169,14 +169,17 @@ class TestCalculatedSummaryContext(TestStandardSummaryContext):
             section_id="default-section", block_id="currency-total-playback-with-fourth"
         )
 
-        context = build_view_context_for_calculated_summary(
+        calculated_summary_context = CalculatedSummaryContext(
             "en",
-            current_location,
             self.schema,
             self.answer_store,
             self.list_store,
             self.progress_store,
             self.metadata,
+        )
+
+        context = calculated_summary_context.build_view_context_for_calculated_summary(
+            current_location
         )
 
         self.check_context(context)
@@ -208,14 +211,18 @@ class TestCalculatedSummaryContext(TestStandardSummaryContext):
 
         skip_answer = Answer("skip-fourth-block-answer", "Yes")
         self.answer_store.add_or_update(skip_answer)
-        context = build_view_context_for_calculated_summary(
+
+        calculated_summary_context = CalculatedSummaryContext(
             "en",
-            current_location,
             self.schema,
             self.answer_store,
             self.list_store,
             self.progress_store,
             self.metadata,
+        )
+
+        context = calculated_summary_context.build_view_context_for_calculated_summary(
+            current_location
         )
 
         self.check_context(context)
@@ -244,14 +251,17 @@ class TestCalculatedSummaryContext(TestStandardSummaryContext):
             section_id="default-section", block_id="unit-total-playback"
         )
 
-        context = build_view_context_for_calculated_summary(
+        calculated_summary_context = CalculatedSummaryContext(
             "cy",
-            current_location,
             self.schema,
             self.answer_store,
             self.list_store,
             self.progress_store,
             self.metadata,
+        )
+
+        context = calculated_summary_context.build_view_context_for_calculated_summary(
+            current_location
         )
 
         self.check_context(context)
@@ -278,14 +288,17 @@ class TestCalculatedSummaryContext(TestStandardSummaryContext):
             section_id="default-section", block_id="percentage-total-playback"
         )
 
-        context = build_view_context_for_calculated_summary(
+        calculated_summary_context = CalculatedSummaryContext(
             "en",
-            current_location,
             self.schema,
             self.answer_store,
             self.list_store,
             self.progress_store,
             self.metadata,
+        )
+
+        context = calculated_summary_context.build_view_context_for_calculated_summary(
+            current_location
         )
 
         self.check_context(context)
@@ -313,14 +326,17 @@ class TestCalculatedSummaryContext(TestStandardSummaryContext):
             section_id="default-section", block_id="number-total-playback"
         )
 
-        context = build_view_context_for_calculated_summary(
+        calculated_summary_context = CalculatedSummaryContext(
             "cy",
-            current_location,
             self.schema,
             self.answer_store,
             self.list_store,
             self.progress_store,
             self.metadata,
+        )
+
+        context = calculated_summary_context.build_view_context_for_calculated_summary(
+            current_location
         )
 
         self.check_context(context)
