+++ tests/integration/questionnaire/test_questionnaire_list_collector.py
@@ -145,7 +145,7 @@ class TestQuestionnaireListCollector(IntegrationTestCase):
 
         self.post({"another-anyone-else": "No"})
 
-        self.assertInBody("Check your answers before continuing")
+        self.assertInBody("List Collector Summary")
 
         self.post()
 
@@ -166,7 +166,7 @@ class TestQuestionnaireListCollector(IntegrationTestCase):
 
         self.post({"another-anyone-else": "No"})
 
-        self.assertInBody("Check your answers before continuing")
+        self.assertInBody("List Collector Summary")
 
         self.post()
